moved: wirechat component to widgets folder
