+++ src/Livewire/New/Group.php
@@ -6,7 +6,7 @@ use Livewire\Attributes\Validate;
 use Livewire\WithFileUploads;
 use Namu\WireChat\Facades\WireChat;
 use Namu\WireChat\Livewire\Concerns\ModalComponent;
-use Namu\WireChat\Livewire\WireChat as WidgetsWireChat;
+use Namu\WireChat\Livewire\Widgets\WireChat as WidgetsWireChat;
 use  Namu\WireChat\Livewire\Concerns\Widget;
 
 class Group extends ModalComponent
