+++ src/Livewire/Widgets/WireChat.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Namu\WireChat\Livewire;
+namespace Namu\WireChat\Livewire\Widgets;
 
 use Illuminate\Contracts\Routing\UrlRoutable;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
@@ -136,6 +136,6 @@ class WireChat extends Component
 
     public function render()
     {
-        return view('wirechat::livewire.wire-chat');
+        return view('wirechat::livewire.widgets.wire-chat');
     }
 }
